// front/src/app/model/Synonyme.model.ts
export class SynonymeModel {
    _id: string;

  default_value: String  ;
  code: String;
  hierarchy_code : String;
  roles:  any [];
  validations : String ;
  required: Boolean;
  variant: Boolean ;
  description_translations: String;
  label: String;
  label_translations: any [];
  
  values : String;
  requirement_level: String;
  values_list :String ;
  type: String ;
  example : String;
  type_parameter: String;
  description :String;


}